refactor: 페이지가 늘어날 것을 대비하여 App.tsx 소스 코드를 pages 폴더 안으로 따로 분리
